imeage classifation project
imeage classifation project in open CV The primary objective of this project is to create a system that can classify images into predefined categories or classes. This is a fundamental task in computer vision and has applications in various domains, including object recognition, medical image analysis, autonomous vehicles, and more.